requirements.txt
django_environ==0.8.0
requests==2.26.0
Django==3.2.8
dj_database_url==0.5.0
django_heroku==0.3.1
djangorestframework==3.12.4
google_api_python_client==2.26.1
django-crispy-forms==1.13.0
django_extensions==2.2.9
django-google-translate==1.1
gunicorn==20.1.0
whitenoise == 5.3.0